requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.3
pymongo==4.6.2
python-dotenv==1.0.1
Werkzeug==3.0.1
bleach==6.1.0 